fixed Start Wizard with V.core Installed
With the changes made in the plugin modules, plus this change we managed to start the plugin if it is pre-installed in the image.
At the moment @persianpros  leave it alone in feeds ... I have to look at a number of things.

adapt screens to the skin and see what possible errors we still have.

The problem of the wizard is that the plugin has its own image restoration module, it wants to interfere with the OV base restoration and there the problems arise.