fixed choosing files by user for backup
There are more index values in the module, but I don't like to assume that it will also be better to remove them.
I have already verified that there are lists with value 0 as in Setup that work perfectly.
So let's hope users have a crash and then we fix it.