Fixed Restore Plugins from Backup and Define Variables OV
Restore with currrentkernelversion is working perfectly thanks persian.
You can add the imageversion, for the restoration, but I am not in favor of inserting it for a simple reason, the platform is in constant progress, as soon as we go up from 10.1 the user will no longer be able to restore the backup and I know that we will have it to explain, with kernel is enough .... and I don't think anyone will make the mistake of going from 7.4 to 10.1 with backup restoration.