Fixed Create SWAP for Flash
@persian, in case you ask in the forum:
swap cannot be created on vfat device, supported devices are ext4, ext3, ext2.

Example: if a user has a stick in vfat format, when creating the SWAP he will only have the option of creating SWAP in Flash, his vfat device will not appear.

I have tested the network mounts, in my case I mount NFS on NAS in / media / net and with this code it is ignored without any problem, that is, there is no option to create them since they do not appear.

With this I think we have this module ready.