BackupManager.py: Define OV Add key Blue for Restore Settings
With this, if the user does not have the possibility to restore a copy by kernel change, he will not have problems, he can use the new blue added key "Restore Settings" for this task, we avoid a tedious task of reconfiguring tuners, channel list etc. Only settings and files chosen by the user are restored, ignoring the opkg package list.

Removed by default the softcams folder, it should not enter the files chosen by the user to save as we use symlink in the module that creates the links to the softcams folder.