Fixed Multiples device on STB for image backup
Keep in mind that Mount Manager handles various mounts on the system, usb2,3, hdd2,3, sdcard, etc. .. that was when problems began to arrive.

With more than one device connected to STB, the module did not manage copies well at all, sometimes using flash because it does not handle the assemblies sent by Mount Manager well.

We do the same better and more stable with less code, this persiampros almost guarantee that we have it fixed, if it fails it must be due to problems with the sticks used by users.