MountManager.py: Specify correct function to red button action use as HDD
Updated the status of messages.
We use device type auto, which automatically detects the device, the line is compact in this way in fstab.
If we use self type device, it has same function for system, with the handicap that does not write correctly in FSTAB, divide the line in two for each device.