Add New Module "Client Mode Box"
Setup a client OV box from a local server with OV image over network.

Based on GBipbox client.

We take advantage of the possibilities that OpenWebif offers us.

The objective:

To be able to configure and have an STB available if we do not have HW resources, either because we do not have an antenna installation resources, or because the box does not have a Tuner or has it broken, it will not be a problem  do not have a storage device, use server harddisk, then let's make the receiver can be operational.

Features:

The module starts with Wizard, it will search for local server if it is available.
once found the client will download the channel list from the local server to have it available.

Setup Resources:

The module has a setup to configure actions:

.-Possibility of synchronizing the server timers, in order to avoid conflicts when trying to record.

.-Be able to synchronize a scheduled download to have the channel list updated in case there were changes on the server.

.-Possibility of having the server harddisk in case we do not have this HW resource on the client.

We have updated the network protocol used.

We have added mounting on hard to avoid losing harddisk from server on client reboot.

Important: to mount on remote hdd reboots, we use the fstab level for automatic mounting, if the server has a NO OV system and it does not have an NFS server mounted, (**and the box has this utility running as a client and use remote HDD from server**) the only thing it will achieve it delays the startup a little more system because OV will search and will not find server.

This is impossible to happen, if the client and the server have OV system...(thanks @persianpros  for add nfs-utils to OV).

It is tested with a server and two types of HW mounted on LNB, unicable and universal and both DVB streams work on the client.

Credits to its original author Sandro Cavazzoni where we have been able to adapt it for OV.