BackupManager.py: fix plugin restore issues.
We had two possible errors, I suppose that depending on the SoC this can be reproduced or not, in my case with HiSi it was evident:
Ensure the elimination of the tmp files, these were maintained despite being in tmp after restarting the box, we ensure their elimination.

Starge5Completed gives errors in the attributes, I want to think that it is because when opkg is not clean in the installation of packages as is the case of the samba package, it generates a first failure in the first restoration, with this it should be solved in a single clean restoration and complete of the installed plugins, we make the restore call directly to stage 5 and 6.