Read BoxInfo and improve user cancellations on backup
The getItem BoxInfo reads are set to make backups.

Enhanced User Cancellations ...
if a copy has an error there is no problem, the mounts disappear from the device and a new backup is attempted.

The problem arises when the copy of a job in progress is canceled, it can be done from the list of extensions with jobTask, here the problem arises:

If trying to make another image copy, the code does not detect any mount and rmtree removes WORKDIR and device is in use (device root is not unmounted in DEV!!!)

Now when this happens, the system will ask you if you want to reboot to remove the media.

Ability to delete any obsolete MAINDESTROOT left on the device as a result of user cancellation. (You can never manually delete TMPDIR and WORKDIR.

We continue with the messagebox system to confirm the deletion of  copy job directories.

All code related to ConfigListEntry is removed, code is read from XML.