Change Method Setup enigma and and improve cancellation of backup
I eliminate inefficient code.
We ensure a root unmount when the device hosting the backup is in use.
I am not given rmtree if the unmount was not done.

We can safely delete old MAINDESTROOT folders, as well as not delete it when there is a backup running.

These folders remain created and were not deleted if the user canceled the backup before it was completely done.

delete code getConfigListEntry now read from setup.XML and use method Setup enigma2